rename `decodeURIComponentSave` to `decodeURIComponentSafe`
